+++ templates/login.html
@@ -4,30 +4,37 @@
         <title>
             Login
         </title>
+
         <meta charset="UTF-8">
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
         <meta name="description" content="이 페이지는 Jinja2와 서버사이드 렌더링에 대한 정보를 제공합니다.">
         <meta name="keywords" content="Jinja2, 서버사이드 렌더링, 웹 개발, 템플릿 엔진">
         <!-- Tailwind CSS CDN -->
         <link href="https://cdn.jsdelivr.net/npm/tailwindcss@2.2.19/dist/tailwind.min.css" rel="stylesheet">
+
+        <!-- jQuery 를 포함합니다. -->
+        <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
+        <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery-cookie/1.4.1/jquery.cookie.min.js"></script>
+
         <script>
             function login() {
-            console.log('로그인 함수')
-            $.ajax({
-                type: "POST",
-                url: "/login",
-                data: { 'email': $('#email').val(), 'password': $('#password').val()},
-                dataType: JSON,
-                success: function (response) {
-                    console.log(response['token'])
-                    $.cookie('mytoken', response['token'], {path:'/'});
+                alert('안녕');
+                $.ajax({
+                    type: "POST",
+                    url: "/login",
+                    data: { 'email': $('#email').val(), 'password': $('#password').val()},
+                    success: function (response) {
+                        if (response['token'] == 'success') {
+                            $.cookie('mytoken', response['token'], { expires: 7, path:'/'});
 
-                    alert('로그인 완료!')
-                    window.location.replace("/main")
-                },
-
-            })
-        }
+                            alert('로그인 완료!')
+                            window.location.replace("/main")
+                        } else {
+                            alert(response['msg'])
+                        }
+                    },
+                });
+            }
         </script>
     </head>
     <body class="bg-gray-100 h-screen flex flex-col justify-center items-center">
